+++ src/renderer/html/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <title>Teston</title>
-    <link rel="stylesheet" href="index.css" />
+    <link rel="stylesheet" href="../css/index.css" />
   </head>
 
   <body>
@@ -17,24 +17,24 @@
         id="test-image-1"
         class="clickable-image"
         data-filename="example.spec.js"
-        src="../images/test-image-1.png"
+        src="../../../images/test-image-1.png"
         alt="test1"
       />
       <img
         id="test-image-2"
         class="clickable-image"
         data-filename="example2.spec.js"
-        src="../images/test-image-2.png"
+        src="../../../images/test-image-2.png"
         alt="test2"
       />
       <img
         id="test-image-3"
         class="clickable-image"
         data-filename="example.spec.js"
-        src="../images/test-image-3.png"
+        src="../../../images/test-image-3.png"
         alt="test3"
       />
     </div>
   </body>
-  <script src="./renderer.js"></script>
+  <script src="../js/index.js"></script>
 </html>
